pipelinev2/requirements.txt
python-dotenv==1.0.0
GitPython==3.1.40
PyYAML==6.0.1
markdown==3.5.1
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.1
tiktoken==0.5.2
openai==1.6.1
langchain-openai==0.0.3
pinecone-client==3.0.1
pytest==7.4.3
nomic==2.0.3
langchain-groq>=0.1.0 